requirements.txt
# Django and Core Dependencies
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-filter==23.5

# Database
psycopg2-binary>=2.9.5
django-extensions==3.2.3

# Cache and Task Queue
redis==5.0.1
celery==5.3.4
django-celery-beat==2.5.0
django-celery-results==2.5.1

# Web Scraping
selenium==4.15.2
beautifulsoup4==4.12.2
requests==2.31.0

# Development Tools
python-decouple==3.8
django-debug-toolbar==4.2.0
drf-yasg==1.21.7

# Production
gunicorn==21.2.0
whitenoise==6.6.0

# Google Cloud Platform
google-cloud-storage==2.10.0
google-cloud-logging==3.8.0
google-cloud-monitoring==2.16.0
google-cloud-secret-manager==2.16.4
django-storages[google]==1.14.2

# Database URL parsing
dj-database-url==2.1.0

# Redis for Django
django-redis==5.4.0

# Utilities
python-dateutil==2.8.2

# Machine Learning and NLP
scikit-learn==1.3.2
numpy==1.24.4
pandas==2.0.3
konlpy==0.6.0
nltk==3.8.1

# Deep Learning and BERT (딥러닝 감성 분석)
torch>=2.0.0
transformers>=4.30.0
tokenizers>=0.13.0
datasets>=2.12.0
accelerate>=0.20.0

# Additional ML Libraries
matplotlib>=3.6.0
seaborn>=0.12.0
jupyter>=1.0.0

# Geographic Information
geopy==2.4.0